Recover unclaimed engine-code bodies: declared-but-undefined and called functions
Recovered bodies called by already-claimed code:
- W8CameraShakeEffect::Evaluate004AE4E0 (GrCycle.cpp): timer-gated
  distance/falloff evaluation; fixes value_08 model from int to float
  (retail call sites push raw floats, Evaluate compares as float).
- TrackItemSpellSource00501D20 (Magic.cpp): equipment/backpack item-spell
  availability bitmap + usage-table counting; retypes g_status unknown_24a0
  as a W8SpellUsageRecord array preserving the 0xC80 span.
- srMatrix3T<float>::RotateAboutZ(double) single-angle overload used by
  W8Monster::UpdateRepresentation; fixes the Monster.cpp call site.

Marker/provenance claims for verified TU-local emissions:
- W8GrowableVector<srVector3T<float>>::Grow (0x4A2670, AnimObj.cpp tail)
- srMatrix3T transposed-transform emission (0x4ED950)
- W8HashTable<unsigned int,int> methods, srArray<srNode*>::setCapacity,
  stMaterial::vInstance, NPC state lifecycle helpers, world-item rebuild,
  character-stat screen controls, material/light destructor emissions

Hygiene: convert earlier-batch C-style casts to static_cast, drop
redundant scalar casts, bool-like-byte fixes (has_alpha_64,
g_flush_pending_603c3a), srQuadWord __int64 reinterpretation documented
with reinterpret-ok.

+++ src/wiz8/engine_code/Spells.cpp
@@ -658,9 +658,8 @@ unsigned char LoadSpellVisualResource004AB580(const W8GrCycleLoadContext* contex
                             break;
                         }
                     }
-                    effect = new W8CameraShakeEffect(
-                        duration, 1, intensity, static_cast<int>(distance * g_world_scale_005ebc40),
-                        0);
+                    effect = new W8CameraShakeEffect(duration, 1, intensity,
+                                                     distance * g_world_scale_005ebc40, 0);
                     if (effect != 0) {
                         effect->cycle_3c = index;
                         effect->frame_40 = frame;
